[DEV-2079] add amount check with user tx
