+++ src/subdomains/core/sell-crypto/process/buy-fiat.controller.ts
@@ -4,8 +4,8 @@ import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { BuyFiat } from './buy-fiat.entity';
-import { BuyFiatService } from './buy-fiat.service';
 import { UpdateBuyFiatDto } from './dto/update-buy-fiat.dto';
+import { BuyFiatService } from './services/buy-fiat.service';
 
 @ApiTags('buyFiat')
 @Controller('buyFiat')
